improv: Add `optional` argument for groups
Optional groups in a config instance are set to `None` if no variables
from the config type used to define the group (with or without default
values) are present in the environment being read from. If any variable
is present, the group reverts to standard rules where all required
variable must be present, but those with defaults are optional. This is
useful for grouping semantically related configuration variables which
should all be ignored if a subset of required ones are not provided.